server/requirements.txt
amqp==5.1.1
appdirs==1.4.4
argcomplete==1.10.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.8.2
billiard==3.6.4.0
black==21.6b0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
compressed-rtf==1.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-admin-interface==0.20.0
django-autoslug==1.9.8
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-countries==7.2.1
django-enum-choices==2.1.4
django-environ==0.9.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jsonfield==1.4.1
django-parler==2.3
django-phonenumber-field==5.2.0
django-redis==5.2.0
django-storages==1.13.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
ffmpeg-python==0.1.17
flake8==3.9.2
future==0.18.2
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
IMAPClient==2.1.0
incremental==22.10.0
isodate==0.6.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
jsoneditor==1.5.1
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.6.1
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.2
olefile==0.46
packaging==21.3
pathspec==0.10.1
pdfminer.six==20191110
phonenumbers==8.12.13
Pillow==8.3.2
priority==1.3.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.3.1
PyJWT==2.1.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
redis==4.2.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
textract==1.6.5
toml==0.10.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.3
youtube-dl==2021.12.17
zope.interface==5.5.0
